Prototype to center the text view
We use the right margin as an indication of the natural size of the
text and add a gutter margin to the left to ensure the text is centered.
(We do not add margin to the right since there the text should be
allowed to exceed the right margin).
In this prototype this feature is always on, if we do it this way we
need to add a boolean property to control it.